add show library
видалив не потрібний файл, логіка готова, потрібні тальки данні із локального сховища підставити
