+++ src/components/show-library/libraryShowList.html
@@ -1,26 +1,11 @@
 <main>
-   <section class="section">
-      <div class="container hero">
-         <ul render="libraryCard" class="library__content"></ul>
-      </div>
-   </section>
+  <section class="section">
+    <div class="container hero">
+      <ul class="hero__list" render="libraryCard"></ul>
+    </div>
+  </section>
 </main>
-<style>
-   .library__content{
-      width: 100%;
-   height: 100%;
-   display: flex;
-   flex-wrap: wrap;
-   gap: 16px;
 
-  /* @media screen and (min-width: 768px) {
-    gap: 30.5px;
-  }
-  @media screen and (min-width: 1280px) {
-    gap: 16px;
-  } */
-   }
-</style>
 <!-- в карточке нужно добавить id, по которому будет добавляться js логика  -->
 <!-- рендер карточки популярных фильмов -->
 <!-- рендер карточки по ключевому слову -->
