+++ src/library.html
@@ -20,7 +20,7 @@
     <include src="./components/show-library/libraryShowList.html"></include>
     <include src=""></include>
     <include src=""></include>
-        <include src="./components/modal-info/modal.html"></include>
+    <include src="./components/modal-info/modal.html"></include>
     <include src="./components/footer/footer.html"></include>
     <include src="./components/trailer/trailer.html"></include>
   </body>
