+++ src/library.js
@@ -1,3 +1,4 @@
-import './components/show-library/queue';
+import './components/show-library/showLibrary';
 
 import './components/trailer/trailer';
+import './components/modal-info/modal';
